LayoutTests/imported/w3c/web-platform-tests/css/css-color/parsing/color-mix-out-of-gamut-expected.txt


FAIL Property color value 'color-mix(in hsl, color(display-p3 0 1 0) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0 0.9764706 0.25882354)
Expected: color(srgb -0.511814 1.01832 -0.310726).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected -0.511814 +/- 0.01, expected -0.511814 but got 0
FAIL Property color value 'color-mix(in hsl, lab(100 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 1 1)
Expected: color(srgb 1.59343 0.58802 1.40564).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59343 +/- 0.01, expected 1.59343 but got 1
FAIL Property color value 'color-mix(in hsl, lab(0 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0.16470589 0 0.13333334)
Expected: color(srgb 0.351376 -0.213938 0.299501).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.351376 +/- 0.01, expected 0.351376 but got 0.16470589
FAIL Property color value 'color-mix(in hsl, lch(100 116 334) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 1 1)
Expected: color(srgb 1.59328 0.588284 1.40527).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59328 +/- 0.01, expected 1.59328 but got 1
FAIL Property color value 'color-mix(in hsl, lch(0 116 334) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0.16470589 0 0.13333334)
Expected: color(srgb 0.351307 -0.213865 0.299236).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.351307 +/- 0.01, expected 0.351307 but got 0.16470589
FAIL Property color value 'color-mix(in hsl, oklab(1 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 0.9764706 1)
Expected: color(srgb 1.59295 0.360371 1.38571).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59295 +/- 0.01, expected 1.59295 but got 1
FAIL Property color value 'color-mix(in hsl, oklab(0 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0 0 0)
Expected: color(srgb 0.0763893 -0.0456266 0.0932598).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.0763893 +/- 0.01, expected 0.0763893 but got 0
FAIL Property color value 'color-mix(in hsl, oklch(1 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 0.9764706 1)
Expected: color(srgb 1.59328 0.358734 1.38664).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59328 +/- 0.01, expected 1.59328 but got 1
FAIL Property color value 'color-mix(in hsl, oklch(0 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0 0 0)
Expected: color(srgb 0.076536 -0.045825 0.0937443).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.076536 +/- 0.01, expected 0.076536 but got 0
FAIL Property color value 'color-mix(in hwb, color(display-p3 0 1 0) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0 0.9764706 0.25882354)
Expected: color(srgb -0.511814 1.01832 -0.310726).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected -0.511814 +/- 0.01, expected -0.511814 but got 0
FAIL Property color value 'color-mix(in hwb, lab(100 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 1 1)
Expected: color(srgb 1.59343 0.58802 1.40564).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59343 +/- 0.01, expected 1.59343 but got 1
FAIL Property color value 'color-mix(in hwb, lab(0 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0.16470589 0 0.13333334)
Expected: color(srgb 0.351376 -0.213938 0.299501).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.351376 +/- 0.01, expected 0.351376 but got 0.16470589
FAIL Property color value 'color-mix(in hwb, lch(100 116 334) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 1 1)
Expected: color(srgb 1.59328 0.588284 1.40527).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59328 +/- 0.01, expected 1.59328 but got 1
FAIL Property color value 'color-mix(in hwb, lch(0 116 334) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0.16470589 0 0.13333334)
Expected: color(srgb 0.351307 -0.213865 0.299236).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.351307 +/- 0.01, expected 0.351307 but got 0.16470589
FAIL Property color value 'color-mix(in hwb, oklab(1 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 0.9764706 1)
Expected: color(srgb 1.59295 0.360371 1.38571).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59295 +/- 0.01, expected 1.59295 but got 1
FAIL Property color value 'color-mix(in hwb, oklab(0 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0 0 0)
Expected: color(srgb 0.0763893 -0.0456266 0.0932598).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.0763893 +/- 0.01, expected 0.0763893 but got 0
FAIL Property color value 'color-mix(in hwb, oklch(1 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 1 0.9764706 1)
Expected: color(srgb 1.59328 0.358736 1.38664).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1.59328 +/- 0.01, expected 1.59328 but got 1
FAIL Property color value 'color-mix(in hwb, oklch(0 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' Colors do not match.
Actual:   color(srgb 0 0 0)
Expected: color(srgb 0.0765361 -0.045825 0.0937443).
Error: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.0765361 +/- 0.01, expected 0.0765361 but got 0
